json / rocketmq serializer test move EncodeHeader into SerializerController
