+++ rocketmq-go/example/consumer_example.go
@@ -20,12 +20,21 @@ import (
 	"fmt"
 	"github.com/apache/incubator-rocketmq-externals/rocketmq-go/model/config"
 	"github.com/apache/incubator-rocketmq-externals/rocketmq-go/service"
+	"github.com/apache/incubator-rocketmq-externals/rocketmq-go/model/constant"
 )
 
 func main() {
+
 	var clienConfig = config.NewClientConfig();
 	clienConfig.SetNameServerAddress("120.55.113.35:9876")
+
+	//use json serializer
 	var mqClient = service.MqClientInit(clienConfig,  nil);
 	fmt.Println(mqClient.TryToFindTopicPublishInfo("GoLang"));
 	//&{false true [{GoLang broker-a 0} {GoLang broker-a 1} {GoLang broker-a 2} {GoLang broker-a 3}] 0xc420016800 0} <nil>
+
+	//use rocketmq serializer
+	constant.USE_HEADER_SERIALIZETYPE = constant.ROCKETMQ_SERIALIZE
+	var mqClient2 = service.MqClientInit(clienConfig,  nil);
+	fmt.Println(mqClient2.TryToFindTopicPublishInfo("GoLang"));
 }
